# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Grandma's Recipes API
  description: API for managing recipes, users, and meal planning
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /users/favorites:
    get:
      tags:
        - Users
      summary: Get user's favorite recipes
      description: Get all recipes marked as favorite by the user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of favorite recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - Users
      summary: Add recipe to favorites
      description: Mark recipe as favorite
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipeId
              properties:
                recipeId:
                  type: string
                  description: Recipe ID to add to favorites
                  example: "716429"
      responses:
        '200':
          description: Recipe added to favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Recipe ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/myRecipes:
    get:
      tags:
        - Users
      summary: Get user's created recipes
      description: Get all recipes created by the user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of user's recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - Users
      summary: Create a new recipe
      description: Create a new user recipe
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreation'
      responses:
        '201':
          description: Recipe created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/familyRecipes:
    get:
      tags:
        - Users
      summary: Get family recipes
      description: Get user's family recipes
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of family recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FamilyRecipe'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - Users
      summary: Add a family recipe
      description: Add a new family recipe
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyRecipeCreation'
      responses:
        '201':
          description: Family recipe created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/last-viewed:
    get:
      tags:
        - Users
      summary: Get last viewed recipes
      description: Get last 3 recipes viewed by user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Last viewed recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    RecipePreview:
      type: object
      properties:
        id:
          type: string
          description: Recipe ID
          example: "716429"
        title:
          type: string
          description: Recipe title
          example: "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs"
        image:
          type: string
          description: Recipe image URL
          example: "https://spoonacular.com/recipeImages/716429-556x370.jpg"
        readyInMinutes:
          type: integer
          description: Preparation time in minutes
          example: 45
        popularity:
          type: integer
          description: Recipe popularity (likes)
          example: 209
        vegetarian:
          type: boolean
          description: Is vegetarian
          example: false
        vegan:
          type: boolean
          description: Is vegan
          example: false
        glutenFree:
          type: boolean
          description: Is gluten free
          example: false
        viewed:
          type: boolean
          description: Has user viewed this recipe
          example: false
        favorite:
          type: boolean
          description: Is recipe in user's favorites
          example: false
...
    RecipeCreation:
      type: object
      required:
        - title
        - image
        - readyInMinutes
        - servings
        - ingredients
        - instructions
      properties:
        title:
          type: string
          description: Recipe title
          example: "Homemade Pizza"
        image:
          type: string
          description: Recipe image URL
          example: "https://example.com/pizza.jpg"
        readyInMinutes:
          type: integer
          description: Preparation time in minutes
          example: 60
        servings:
          type: integer
          description: Number of servings
          example: 4
        vegetarian:
          type: boolean
          description: Is vegetarian
          example: false
        vegan:
          type: boolean
          description: Is vegan
          example: false
        glutenFree:
          type: boolean
          description: Is gluten free
          example: false
        ingredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "flour"
              amount:
                type: number
                example: 2
              unit:
                type: string
                example: "cups"
        instructions:
          type: string
          description: Cooking instructions
          example: "1. Mix ingredients..."
    FamilyRecipe:
      type: object
      properties:
        id:
          type: string
          description: Recipe ID
          example: "family_1"
        title:
          type: string
          description: Recipe title
          example: "Grandma's Chicken Soup"
        image:
          type: string
          description: Recipe image URL
          example: "https://example.com/soup.jpg"
        author:
          type: string
          description: Who created this recipe
          example: "Grandma Sarah"
        whenToMake:
          type: string
          description: When is this recipe traditionally made
          example: "Friday nights and holidays"
        ingredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "chicken"
              amount:
                type: number
                example: 1
              unit:
                type: string
                example: "whole"
        instructions:
          type: string
          description: Cooking instructions
          example: "1. Clean the chicken..."
    FamilyRecipeCreation:
      type: object
      required:
        - title
        - author
        - whenToMake
        - ingredients
        - instructions
      properties:
        title:
          type: string
          description: Recipe title
          example: "Grandma's Chicken Soup"
        image:
          type: string
          description: Recipe image URL (optional)
          example: "https://example.com/soup.jpg"
        author:
          type: string
          description: Who created this recipe
          example: "Grandma Sarah"
        whenToMake:
          type: string
          description: When is this recipe traditionally made
          example: "Friday nights and holidays"
        ingredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "chicken"
              amount:
                type: number
                example: 1
              unit:
                type: string
                example: "whole"
        instructions:
          type: string
          description: Cooking instructions
          example: "1. Clean the chicken..."
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operation successful"
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Error occurred"
        error:
          type: string
          example: "Invalid input"
    RecipeResponse:
      type: object
      properties:
        message:
          type: string
          example: "Recipe created successfully"
        recipeId:
          type: string
          example: "123456"